# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: SAE Project BUT2 API
  description: |
    API REST pour notre application web "Les Terres Du Lions".

    ## Fonctionnalités
    - Authentification et gestion des utilisateurs avec (JWT)
    - Gestion des locations (points d'intérêt)
    - Catalogue de produits 
    - Événements et réservations
    - Système de quêtes gamifié
    - Blog et articles
    - QR Codes
    
    ## Authentification
    Les routes protégées nécessitent un token JWT dans le header `Authorization: Bearer <token>`.
servers:
  - url: http://localhost:3000/api/v1
    description: Serveur de développement local
...
paths:
...
  /events:
    get:
      tags: [Events]
      summary: Lister tous les événements
      operationId: getEvents
      responses:
        '200':
          description: Liste des événements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      tags: [Events]
      summary: Créer un événement
      operationId: createEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '201':
          description: Événement créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
  /events/{id}:
    get:
      tags: [Events]
      summary: Récupérer un événement par ID
      operationId: getEventById
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Événement trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Events]
      summary: Mettre à jour un événement
      operationId: updateEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '200':
          description: Événement mis à jour
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Events]
      summary: Supprimer un événement
      operationId: deleteEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Événement supprimé
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /events/book:
    post:
      tags: [Events, Non-Trivial]
      summary: Réserver un événement
      description: |
        **Route Non-Triviale** : Gère la réservation d'événements avec validation de capacité.
        
        **Complexité** :
        - Vérification de disponibilité (sold + quantity ≤ capacity)
        - Calcul du prix total (price × quantity)
        - Transaction atomique (création réservation + mise à jour sold)
        - Gestion d'erreurs métier (capacité dépassée)
        
        **Tables impliquées** : Events, EventReservation, Users
      operationId: bookEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingInput'
      responses:
        '201':
          description: Réservation créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventReservation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /events/user/reservations:
    get:
      tags: [Events, Non-Trivial]
      summary: Récupérer mes réservations
      description: |
        **Route Non-Triviale** : Récupère toutes les réservations d'événements d'un utilisateur.
        
        **Complexité** :
        - Jointure triple : User → EventReservation → Events
        - Filtrage par utilisateur authentifié (JWT)
        - Inclusion des détails d'événements liés
        
        **Tables impliquées** : EventReservation, Events, Users
      operationId: getUserReservations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des réservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventReservation'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ========== QUESTS ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/login
  parameters:
...
    EventId:
      name: id
      in: path
      required: true
      description: Identifiant unique de l'événement
      schema:
        type: integer
        minimum: 1
      example: 1
...
  schemas:
...
    Event:
      type: object
      properties:
        id_event:
          type: integer
          example: 1
        title:
          type: string
          example: Festival des Lumières
        description:
          type: string
          nullable: true
          example: Un festival magique illuminant la nuit
        start_time:
          type: string
          format: date-time
          example: '2025-01-15T18:00:00Z'
        end_time:
          type: string
          format: date-time
          example: '2025-01-15T23:00:00Z'
        price:
          type: number
          example: 15.00
        capacity:
          type: integer
          example: 200
        sold:
          type: integer
          example: 45
        id_location:
          type: integer
          example: 1
    EventInput:
      type: object
      required: [title, start_time, end_time, price, capacity, id_location]
      properties:
        title:
          type: string
          example: Festival des Lumières
        description:
          type: string
          example: Un festival magique illuminant la nuit
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        price:
          type: number
          example: 15.00
        capacity:
          type: integer
          example: 200
        id_location:
          type: integer
          example: 1
    BookingInput:
      type: object
      required: [id_event, quantity]
      properties:
        id_event:
          type: integer
          example: 1
        quantity:
          type: integer
          minimum: 1
          example: 2
    EventReservation:
      type: object
      properties:
        id_reservation:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        total_price:
          type: number
          example: 30.00
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED]
          example: CONFIRMED
        created_at:
          type: string
          format: date-time
        id_user:
          type: integer
        id_event:
          type: integer

    # ===== QUEST SCHEMAS =====
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Une erreur est survenue
        message:
          type: string
          example: Message d'erreur détaillé
  responses:
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Ressource non trouvée
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Données de requête invalides
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Non autorisé
...